address misc pr feedback
